requirements.txt
fastapi
uvicorn
vertexai
google-cloud-discoveryengine